requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
matplotlib
streamlit
requests
Pillow
joblib